Split Database and Collection functional tests
