+++ tests/Database/DatabaseFunctionalTest.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace MongoDB\Tests\Database;
+
+/**
+ * Functional tests for the Database class.
+ */
+class DatabaseFunctionalTest extends FunctionalTestCase
+{
+    public function testDrop()
+    {
+        $writeResult = $this->manager->executeInsert($this->getNamespace(), array('x' => 1));
+        $this->assertEquals(1, $writeResult->getInsertedCount());
+
+        $commandResult = $this->database->drop();
+        $this->assertCommandSucceeded($commandResult);
+        $this->assertCollectionCount($this->getNamespace(), 0);
+    }
+}
